==> LIMPIEZA/1_Pair_Prog_Limpieza_1.ipynb <==
# 03/05 - Pair Programming Limpieza 1 - Guadalupe & Lara

import pandas as pd

df = pd.read_csv("union_ataques.csv", index_col = 0)
df.head(2)

df2 = df.copy()

columnas = {col : col.replace(" ", "_").lower() for col in df2.columns}

df2.rename(columns = columnas, inplace = True)

df2.head(2)

df2.columns

df2.drop(["href", "unnamed:_22", "unnamed:_23", "case_number.1", "case_number.2", "original_order", "investigator_or_source", "pdf", "href_formula"], axis = 1, inplace = True)


Al empezar a trabajar con este dataframe ya eliminamos algunas columnas que a priori no nos interesaban. Ahora llega el momento de eliminar alguna columna más. En este caso vamos tendréis que eliminar las columnas que no nos sean útiles para contestar a nuestras preguntas. Pero ojo ⚠️, haced una copia del dataframe para no "cargarnos" el dataframe original y perder la info.

# Ejercicio 1

¿Hay valores duplicados en nuestro dataframe? En caso de que los haya, eliminándolos.


filas_duplicadas = df2.duplicated().sum()
filas_duplicadas

No tenemos valores duplicados en nuestro dataframe. ✌️ 

En caso de que quisiéramos eliminar algún registro duplicado utilizaríamos df.drop.duplicates(subsets = ["lista_columnas"])

# Ejercicio 2
Como hemos visto, algunas columnas no tienen el tipo de datos que deberían. Cambiad el tipo de dato para la columna de year.


df2.dtypes

#Cambiamos "year" a tipo de dato fecha mediante astype
df2["year"].astype("datetime64", errors = "ignore", copy = True)
df2.dtypes

#Cambiamos a tipo fecha con el método "to_datetime" y sigue siendo "float64"
pd.to_datetime(df2["year"])
df2.dtypes

👩🏽‍🦰👩🏽‍🦱💬      

**Cambiamos "year" a tipo int como en la teoría**
- Nos quedamos con esta versión que creemos que es la correcta y que muestra el año como un int (esto puede facilitar hacer operaciones)

df2["year"] = df2["year"].astype('Int64', errors = "ignore") 
df2.dtypes

# Ejercicio 3
En la columna de country poner todos los valores en minúsculas.
💡 Pista Tendréis que usar una función o una lambda 💡


df2["country"].dtypes

df2["country"].isna().sum()

Como en la celda anterior hemos comprobado que hay NaN, nuestra función llevará un try/except para evitar el error por NaN:

def minusculizator (columna):
    try:
        columna = columna.lower()
    except:
        pass
    return columna

df2["country"] = df2["country"].apply(minusculizator)

df2["country"] 

# Ejercicio 4
Guardamos el csv para seguir trabajando en el siguiente ejercicio de pair de limpieza.

df2.to_csv("1.ataques.csv")
