==> LIMPIEZA/6_Pair_Encoding.ipynb <==
# 11/05 - Pair Programming Limpieza 6 - Guadalupe & Lara

import pandas as pd
import re
import numpy as np
from sklearn.preprocessing import LabelEncoder 
from sklearn.preprocessing import OneHotEncoder 


df = pd.read_csv("attacks_limpiando_con_coordenadas.csv", index_col = 0)

En la sesión de hoy trabajaremos con la codificación de las variables categóricas. Recordemos que esto es esencial ya que muchos modelos de machine learning no trabajan con variables numéricas, y debemos convertirlas a variables numéricas.

Ninguna de nuestras columnas categóricas en este dataframe parece tener un orden por lo que descartamos usar la codificación de tipo Ordinal Encoding. Trabajaremos, por lo tanto, con codificación de tipo **One Hot Encoding y Label Encoding**.

Lo primero que debemos hacer es identificar cuáles de nuestras columnas es categóricas. Extraed el nombre de las columnas que contienen variables categóricas.


df_cat = df.select_dtypes(include = "object")
columnas_cat = df_cat.columns
columnas_cat 

Si nos fijamos en la columna de country es my complicado codificar esa columna ya que hay muchísimas categorías. Para ello, vamos a usar una librería que llamada geopy que nos devuelve las coordenadas de un pais, ciudad, pueblo, lo que imaginemos! El objetivo de esto es poder clasificar nuestros paises en Hemisferio Norte y Sur para poder hacer una codificación más eficaz. Os dejamos por aquí la documentación de la librería para que veais como funciona y el código:

#def extraer_coordenadas(x):
#   try:
#        geolocator = Nominatim(user_agent = "ana")
#        location = geolocator.geocode(x)
#        return location.lat, location.long
#    except:
#        return "no hay datos" 

Nuestra recomendación... no la ejecuteis, tarda muchísimo en aplicarse a nuestro dataframe. Os dejamos el dataframe con las coordenadas al final de la explicación de la sesión.

**2. Codificar "country"**
- Como ya tenemos las coordenadas, deberéis codificar la columna en función de si pertenecen al Hemisferio Norte (HN) o Sur (HS). Aquellas que pertencezcan al HN le pondremos un 1 y a las del HS le pondremos un 0.
- Nota 📌  Perteneceran al HN las que su latitud sea positiva y las del HS las que su latitud sea negativa.



negativos = df["latitud"].str.contains("-")

# Primero creamos una nueva columna llamada "hemisferio" con pandas, para dividir en función de HS o HN
# Después, codificamos la columna con label_encoding para que HS = 0, y HN = 1

patron_neg = r".*\-[0-9].*"
patron_pos = r".*[0-9].*"
 
def hemispherificator(col):
    try:
        if re.findall(patron_neg, col):
            return "HS"
        elif re.findall(patron_pos, col):
            return "HN"
        else:
            return np.nan
    except:
        np.nan

df["Hemisferio"] = df["latitud"].apply(hemispherificator)
df

👩🏽‍🦰👩🏽‍🦱💬      
- Hemos optado por convertir los valores de "latitud" que no eran "correctos" (dígitos, con o sin el "-") en np.nan. 
- Entonces, al hacer el encoding tendremos 3 valores numéricos (0, 1 y 2). En un primer intento hacemos "label encoding" pero los valores que otorga a HM y HN, al ser aleatorios, no son los que queremos por el enunciado (vaya, que nos pone HS = 1 y HN = 0 y lo queremeos alrevés). Entonces, lo cambiamos por un map para poner en el diccionario el valor a nuestro gusto.

mapa = {"HN": 1, "HS" : 0, np.nan : np.nan} 

df["Hemisferio_encoded"] = df["Hemisferio"].map(mapa)
df.head(10)

df["Hemisferio"].unique()

**3. Aplicad un get_dummies a las columnas de fatal y age.**


columnas = ["age", "fatal"]

for col in df[columnas]:
    dummies = pd.get_dummies(df[col], prefix_sep = "_", prefix = col, dtype = int)
    df = pd.concat([df , dummies], axis = 1)

df.columns

**4. Aplicad un Label Encoding a la columna species , fecha y type.**


le = LabelEncoder()

lista_cols = ["species_", "fecha_limpia", "type"]

for col in df[lista_cols]:
    df[col]=le.fit_transform(df[col])

df.head()


df.columns

**5. Guardad los cambios en un csv nuevo para seguir trabajando con él en siguientes ejercicios de pair.**

df.to_csv("2_ataques_coordenadas_pair6")
